Reorganised classes relate to the AutoFix feature
