+++ src/main/java/io/meterian/jenkins/glue/MeterianPlugin.java
@@ -10,8 +10,7 @@ import hudson.model.FreeStyleProject;
 import hudson.tasks.BuildStepDescriptor;
 import hudson.tasks.Builder;
 import hudson.util.FormValidation;
-import io.meterian.jenkins.ClientRunner;
-import io.meterian.jenkins.PullRequestCreator;
+import io.meterian.jenkins.autofixfeature.PullRequestCreator;
 import io.meterian.jenkins.core.Meterian;
 import io.meterian.jenkins.io.HttpClientFactory;
 import net.sf.json.JSONObject;
