+++ src/main/java/io/meterian/jenkins/glue/MeterianStep.java
@@ -3,8 +3,7 @@ package io.meterian.jenkins.glue;
 import hudson.EnvVars;
 import hudson.Extension;
 import hudson.model.TaskListener;
-import io.meterian.jenkins.ClientRunner;
-import io.meterian.jenkins.PullRequestCreator;
+import io.meterian.jenkins.autofixfeature.PullRequestCreator;
 import io.meterian.jenkins.core.Meterian;
 import io.meterian.jenkins.glue.executors.GerritExecutor;
 import io.meterian.jenkins.glue.executors.MeterianExecutor;
